Remove use of reference in Spi, I2C, Gpio structs
When these structs use a reference to FtInner (or an Arc<Mutex<FtInner>>) it makes it impossible to have both, the Device and the SPI struct inside a reference (due to self-referential structs) except for environments with static lifetimes.
Storing an Arc<Mutex<FtInner>> instead of &Arc<Mutex<FtInner>> leads to an additional clone when creating these structs, but removes the above-mentioned limitations.